requirements.txt
gradio>=4.0.0
supabase>=2.0.0
httpx>=0.25.0
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
